requirements.txt
streamlit==1.50.0
pydantic==2.11.9
pydantic-ai==1.0.11
PyPDF2==3.0.1
python-docx==1.2.0
